Improve auth layout
