+++ frontend/src/app/(auth)/common/UnauthPage.tsx
@@ -19,10 +19,11 @@ const UnauthPage = () => {
   }
 
   return (!user &&
-    <div className="max-w-3xl p-4 mx-auto">
-      <h1 className="text-white text-h1 font-bold">Oops...</h1>
+    <div className="max-w-3xl p-4 mx-auto mt-20">
+      <h1 className="text-secondary text-h1 font-bold">Oops...</h1>
+      <p className="text-primary-300 my-5">You are unauthorized to view this page.</p>
 
-      <button onClick={handleLogin} className="">You are unauthorized to view this page.</button>
+      <button onClick={handleLogin} className="bg-yellow-500 hover:bg-yellow-300 px-4 py-2 rounded-lg">Login</button>
     </div>
   );
 };
