+++ frontend/src/app/(auth)/home/components/DashboardCard.tsx
