+++ frontend/src/app/(auth)/home/page.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import { useAuth } from "@/components/auth/AuthContext";
+import AuthDashboard from "./pages/AuthDashboard";
+import UnauthPage from "../common/UnauthPage";
+
+const Dashboard = () => {
+  const { user } = useAuth();
+
+  return (
+    user ? <AuthDashboard/> : <UnauthPage/>
+  );
+};
+
+export default Dashboard;
