+++ frontend/src/app/(auth)/home/pages/LandingPage.tsx
@@ -1,0 +1,9 @@
+const LandingPage = () => {
+  return (
+    <div>
+      <h1>Landing Page</h1>
+    </div>
+  );
+}
+
+export default LandingPage;
