+++ frontend/src/app/(auth)/layouts/AuthLayout.tsx
